+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## 2017-1X-XX FreshRSS 1.8.1-dev
 
+* UI
+	Show API subscription URL [#1247](https://github.com/FreshRSS/FreshRSS/issues/1247)
 * SimplePie
 	* Remove "SimplePie" name from HTTP User-Agent string [#1656](https://github.com/FreshRSS/FreshRSS/pull/1656)
 * Bug fixing
 	* Work-around for `CURLOPT_FOLLOWLOCATION` `open_basedir` bug in favicons and PubSubHubbub [#1655](https://github.com/FreshRSS/FreshRSS/issues/1655)
 * Misc.
-	* Travis translation validation tool [#1653](https://github.com/FreshRSS/FreshRSS/pull/1653)
+	* Translation validation tool [#1653](https://github.com/FreshRSS/FreshRSS/pull/1653)
+	* Translation manipulation tool [#1658](https://github.com/FreshRSS/FreshRSS/pull/1658)
 
 
 ## 2017-10-01 FreshRSS 1.8.0